+++ pig-upms/pig-upms-api/src/main/java/com/pig4cloud/pig/admin/api/entity/AppEstateEntity.java
@@ -25,7 +25,7 @@ public class AppEstateEntity extends Model<AppEstateEntity> {
 	*/
     @TableId(type = IdType.ASSIGN_ID)
     @Schema(description="建筑ID")
-    private Integer estateId;
+    private Long estateId;
 
 	/**
 	* 建筑名称
