+++ pig-upms/pig-upms-api/src/main/java/com/pig4cloud/pig/admin/api/vo/EstateSuiteContractVO.java
@@ -45,7 +45,7 @@ public class EstateSuiteContractVO implements Serializable {
 	 */
 	@TableId(type = IdType.ASSIGN_ID)
 	@Schema(description="建筑ID")
-	private Integer estateId;
+	private Long estateId;
 
 	/**
 	 * 建筑名称
