+++ pig-upms/pig-upms-api/src/main/java/com/pig4cloud/pig/admin/api/vo/EstateSuiteVO.java
@@ -47,7 +47,7 @@ public class EstateSuiteVO implements Serializable {
 	 */
 	@TableId(type = IdType.ASSIGN_ID)
 	@Schema(description="建筑ID")
-	private Integer estateId;
+	private Long estateId;
 
 	/**
 	 * 建筑名称
