+++ pig-upms/pig-upms-biz/src/main/java/com/pig4cloud/pig/admin/mapper/AppContractMapper.java
@@ -14,6 +14,6 @@ public interface AppContractMapper extends BaseMapper<AppContractEntity> {
 	 * @param suiteId
 	 * @return
 	 */
-	List<AppSuiteMapper> listContractsBySuiteId(Integer suiteId);
+	List<AppSuiteMapper> listContractsBySuiteId(Long suiteId);
 
 }
